refactor: group WI15 calibration procedures
