+++ omotion/calibration/__init__.py
@@ -1,0 +1,6 @@
+"""Auditable calibration procedure workflows for OpenMOTION devices.
+
+The package separates pure acceptance rules, UI-neutral workflows, hardware
+adapters, and report rendering. Operator scripts and future application buttons
+should depend on these modules rather than duplicating procedure logic.
+"""
